refactor(renderer): extract local system command runner
